drivers: pinctrl: renesas: rcar: Change pfc_rcar_get_reg_index() prototype
Replace uint8_t with uint16_t for pin parameter. Some SoCs may have
more than 8 GPIO groups.

+++ drivers/pinctrl/renesas/rcar/pfc_r8a77951.c
@@ -538,7 +538,7 @@ const struct pfc_drive_reg *pfc_rcar_get_drive_regs(void)
 	return pfc_drive_regs;
 }
 
-int pfc_rcar_get_reg_index(uint8_t pin, uint8_t *reg_index)
+int pfc_rcar_get_reg_index(uint16_t pin, uint8_t *reg_index)
 {
 	/* There is only one register on Gen 3 */
 	*reg_index = 0;
